\name{agdex.res}
\alias{agdex.res}
\title{Saved result returned by agdex()}
\description{\emph{agdex.res} is result object returned by \emph{agdex} function. 
We saved the result so that we do not have to repeat calling \emph{agdex} to show the examples 
 in the documentation files (for visualization, writing output files etc.).
Users may not need to load this result data as long as the returned result of \emph{agdex} is not deleted from user's current R workspace. 
}
\usage{
data(agdex.res)
}
\value{Components of \emph{agdex.res} have the same meaning as the result object returned by function \emph{agdex()}}

\seealso{
\code{\link{agdex}}
}
